# src/main/resources/swagger/InsuredProcess.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "InsuredProcess API",
    "description": "API documentation for InsuredProcess application",
    "contact": {
      "name": "API Support",
      "email": "support@example.org"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/product/update": {
      "put": {
        "tags": [
          "product-controller"
        ],
        "operationId": "updateProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/product/buy": {
      "post": {
        "tags": [
          "product-controller"
        ],
        "operationId": "buyProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductBuyNewRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/clients/create": {
      "post": {
        "tags": [
          "client-controller"
        ],
        "operationId": "createNewClient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientAuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "uniqueItems": true,
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/clients/client/existing": {
      "post": {
        "tags": [
          "client-controller"
        ],
        "operationId": "existingClient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientAuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductInfo"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UpdateProductRequestDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "newName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "ProductBuyNewRequestDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          }
        }
      },
      "ClientAuthRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "contactMethod": {
            "$ref": "#/components/schemas/ContactMethod"
          }
        }
      },
      "ContactMethod": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "format": "int64"
          },
          "methodType": {
            "type": "string"
          },
          "methodValue": {
            "type": "string"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "category": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "ProductInfo": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "productDescription": {
            "type": "string"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          }
        }
      }
    }
  }
}